# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backcast EVS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/wbes": {
      "get": {
        "tags": [
          "WBEs"
        ],
        "summary": "Read Wbes",
        "description": "Retrieve WBEs with server-side search, filtering, and sorting.\n\nSupports two modes:\n1. **Hierarchical filtering** (project_id/parent_wbe_id): Returns list without pagination\n2. **General listing** (no hierarchical filters): Returns paginated response with search/filter/sort\n\nHierarchical Filtering:\n- project_id only: All WBEs in project\n- project_id + parent_wbe_id: Child WBEs of specified parent\n- parent_wbe_id='null': Root WBEs (parent_wbe_id IS NULL)\n\nGeneral Listing (when no hierarchical filters):\n- **Search**: Case-insensitive search across code and name\n- **Filters**: Filter by level, code, name (format: \"column:value;column:value1,value2\")\n- **Sorting**: Sort by any field (asc/desc)\n- **Pagination**: Returns total count for proper pagination UI\n- **Mode**: Branch mode - \"merged\" (combine with main) or \"isolated\" (current branch only)\n\nRequires read permission.",
        "operationId": "get_wbes",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by project ID",
              "title": "Project Id"
            },
            "description": "Filter by project ID"
          },
          {
            "name": "parent_wbe_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by parent WBE ID (use 'null' string for root WBEs)",
              "title": "Parent Wbe Id"
            },
            "description": "Filter by parent WBE ID (use 'null' string for root WBEs)"
          },
          {
            "name": "branch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Branch name",
              "default": "main",
              "title": "Branch"
            },
            "description": "Branch name"
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(merged|isolated)$",
              "description": "Branch mode: merged (combine with main) or isolated (current branch only)",
              "default": "merged",
              "title": "Mode"
            },
            "description": "Branch mode: merged (combine with main) or isolated (current branch only)"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Search term (code, name)",
              "title": "Search"
            },
            "description": "Search term (code, name)"
          },
          {
            "name": "filters",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filters in format 'column:value;column:value1,value2'",
              "title": "Filters"
            },
            "description": "Filters in format 'column:value;column:value1,value2'"
          },
          {
            "name": "sort_field",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Field to sort by",
              "title": "Sort Field"
            },
            "description": "Field to sort by"
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(asc|desc)$",
              "description": "Sort order (asc or desc)",
              "default": "asc",
              "title": "Sort Order"
            },
            "description": "Sort order (asc or desc)"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: get WBEs as of this timestamp (ISO 8601)",
              "title": "As Of"
            },
            "description": "Time travel: get WBEs as of this timestamp (ISO 8601)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WBEs"
        ],
        "summary": "Create Wbe",
        "description": "Create a new WBE. Requires create permission.",
        "operationId": "create_wbe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WBECreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WBERead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wbes/{wbe_id}": {
      "get": {
        "tags": [
          "WBEs"
        ],
        "summary": "Read Wbe",
        "description": "Get a specific WBE by id. Requires read permission.\n\nSupports time-travel queries via the as_of parameter to view\nthe WBE's state at any historical point in time.",
        "operationId": "get_wbe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "wbe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Wbe Id"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Branch name",
              "default": "main",
              "title": "Branch"
            },
            "description": "Branch name"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: get WBE state as of this timestamp (ISO 8601)",
              "title": "As Of"
            },
            "description": "Time travel: get WBE state as of this timestamp (ISO 8601)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WBERead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "WBEs"
        ],
        "summary": "Update Wbe",
        "description": "Update a WBE. Requires update permission.",
        "operationId": "update_wbe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "wbe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Wbe Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WBEUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WBERead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "WBEs"
        ],
        "summary": "Delete Wbe",
        "description": "Soft delete a WBE. Requires delete permission.",
        "operationId": "delete_wbe",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "wbe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Wbe Id"
            }
          },
          {
            "name": "control_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional control date for deletion",
              "title": "Control Date"
            },
            "description": "Optional control date for deletion"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wbes/{wbe_id}/breadcrumb": {
      "get": {
        "tags": [
          "WBEs"
        ],
        "summary": "Read Wbe Breadcrumb",
        "description": "Get breadcrumb trail for a WBE (project + ancestor path). Requires read permission.",
        "operationId": "get_wbe_breadcrumb",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "wbe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Wbe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WBEBreadcrumb"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/wbes/{wbe_id}/history": {
      "get": {
        "tags": [
          "WBEs"
        ],
        "summary": "Read Wbe History",
        "description": "Get version history for a WBE. Requires read permission.",
        "operationId": "get_wbe_history",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "wbe_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Wbe Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WBERead"
                  },
                  "title": "Response Get Wbe History"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WBEBreadcrumb": {
        "properties": {
          "project": {
            "$ref": "#/components/schemas/ProjectBreadcrumbItem"
          },
          "wbe_path": {
            "items": {
              "$ref": "#/components/schemas/WBEBreadcrumbItem"
            },
            "type": "array",
            "title": "Wbe Path",
            "description": "Ordered list of WBEs from root to current (last item is the current WBE)"
          }
        },
        "type": "object",
        "required": [
          "project",
          "wbe_path"
        ],
        "title": "WBEBreadcrumb",
        "description": "Breadcrumb trail for a WBE showing project and ancestor path."
      },
...
      "WBECreate": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id",
            "description": "Parent project root ID"
          },
          "code": {
            "type": "string",
            "maxLength": 50,
            "title": "Code",
            "description": "WBS code (e.g., 1.2.3)"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name",
            "description": "WBE name"
          },
          "budget_allocation": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Budget Allocation",
            "description": "Budget allocation",
            "default": "0"
          },
          "level": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Level",
            "description": "Hierarchy level",
            "default": 1
          },
          "parent_wbe_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Wbe Id",
            "description": "Parent WBE root ID"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Description"
          },
          "wbe_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wbe Id",
            "description": "Root WBE ID (internal use only for seeding)"
          },
          "branch": {
            "type": "string",
            "title": "Branch",
            "description": "Branch name for creation (defaults to main if not specified)",
            "default": "main"
          },
          "control_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Control Date",
            "description": "Optional control date for creation (valid_time start)"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "code",
          "name"
        ],
        "title": "WBECreate",
        "description": "Schema for creating a new WBE."
      },
      "WBERead": {
        "properties": {
          "project_id": {
            "type": "string",
            "format": "uuid",
            "title": "Project Id",
            "description": "Parent project root ID"
          },
          "code": {
            "type": "string",
            "maxLength": 50,
            "title": "Code",
            "description": "WBS code (e.g., 1.2.3)"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "title": "Name",
            "description": "WBE name"
          },
          "budget_allocation": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Budget Allocation",
            "description": "Budget allocation",
            "default": "0"
          },
          "level": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Level",
            "description": "Hierarchy level",
            "default": 1
          },
          "parent_wbe_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Wbe Id",
            "description": "Parent WBE root ID"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 5000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Description"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "wbe_id": {
            "type": "string",
            "format": "uuid",
            "title": "Wbe Id"
          },
          "branch": {
            "type": "string",
            "title": "Branch"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By Name"
          },
          "parent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Name"
          },
          "deleted_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted By"
          },
          "valid_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid Time"
          },
          "transaction_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Time"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "code",
          "name",
          "id",
          "wbe_id",
          "branch",
          "created_by"
        ],
        "title": "WBERead",
        "description": "Schema for reading WBE data."
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}